# applauds/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API",
    "description": ""
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/create": {
      "post": {
        "tags": [
          "Applauds"
        ],
        "description": "To save applaud details",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "postId": {
                  "example": "any"
                },
                "applaudKey": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
    "/update": {
      "put": {
        "tags": [
          "Applauds"
        ],
        "description": "To update applaud details",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "applaudId": {
                  "example": "any"
                },
                "applaudKey": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
    "/delete/{id}": {
      "delete": {
        "tags": [
          "Applauds"
        ],
        "description": "To delete applaud details",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Enter Applaud ID"
          }
        ],
        "responses": {}
      }
    },
    "/fetch/{postId}": {
      "get": {
        "tags": [
          "Applauds"
        ],
        "description": "To fetch all applaud details of a post",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Enter Post ID"
          }
        ],
        "responses": {}
      }
    }
  }
}